Move projet setup script to individual file.
This allows the file to be exectued in a different project.
